live_dashboard_lite.py
#!/usr/bin/env python3
"""
ARBTRONX Live Trading Dashboard - Lite Version
Works without pandas/numpy/ccxt for initial deployment
"""

import os
import time
import asyncio
from fastapi import FastAPI
from fastapi.responses import HTMLResponse
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Create FastAPI app
app = FastAPI(
    title="ARBTRONX Live Trading Dashboard",
    description="Professional cryptocurrency trading platform",
    version="1.0.0"
)

# Add CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Global variables for system status
system_status = {
    "api_connected": False,
    "trading_active": False,
    "last_update": time.time()
}

@app.on_event("startup")
async def startup_event():
    """Start the dashboard immediately"""
    print("🚀 Starting ARBTRONX Live Dashboard (Lite Version)...")
    print("✅ Health endpoints ready")
    print("📊 Dashboard available immediately")
    print("🔄 Full trading system will be added incrementally...")
    
    # Update system status
    system_status["last_update"] = time.time()
    
    print("✅ ARBTRONX Dashboard startup complete!")

@app.get("/health")
async def health_check():
    """Health check endpoint for deployment platforms"""
    return {
        "status": "healthy",
        "service": "ARBTRONX Live Trading Dashboard",
        "version": "1.0.0",
        "api_connected": system_status["api_connected"],
        "timestamp": time.time()
    }

@app.get("/status")
async def status_check():
    """Backup health check endpoint"""
    return {"status": "ok", "service": "ARBTRONX", "message": "Dashboard is running"}

@app.get("/", response_class=HTMLResponse)
async def dashboard():
    """Serve the Live Trading Dashboard"""
    
    # Get environment variables
    api_key = os.getenv('BINANCE_API_KEY', 'Not configured')
    user_name = os.getenv('USER_NAME', 'User')
    user_email = os.getenv('USER_EMAIL', 'user@example.com')
    user_id = os.getenv('USER_ID', 'unknown')
    
    # Mask API key for security
    masked_key = f"{api_key[:8]}...{api_key[-8:]}" if len(api_key) > 16 else "Not configured"
    
    html_content = f"""
    <!DOCTYPE html>
    <html lang="en">
    <head>
        <meta charset="UTF-8">
        <meta name="viewport" content="width=device-width, initial-scale=1.0">
        <title>ARBTRONX Live Trading Dashboard</title>
        <style>
            * {{
                margin: 0;
                padding: 0;
                box-sizing: border-box;
            }}
            
            body {{
                font-family: 'Segoe UI', Tahoma, Geneva, Verdana, sans-serif;
                background: linear-gradient(135deg, #0a0a0a 0%, #1a1a2e 50%, #16213e 100%);
                color: #00ff88;
                min-height: 100vh;
                padding: 20px;
            }}
            
            .container {{
                max-width: 1200px;
                margin: 0 auto;
            }}
            
            .header {{
                text-align: center;
                margin-bottom: 40px;
            }}
            
            .logo {{
                font-size: 3rem;
                font-weight: bold;
                margin-bottom: 10px;
                text-shadow: 0 0 20px #00ff88;
            }}
            
            .subtitle {{
                font-size: 1.2rem;
                color: #88ffaa;
                margin-bottom: 20px;
            }}
            
            .status-grid {{
                display: grid;
                grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
                gap: 20px;
                margin-bottom: 30px;
            }}
            
            .status-card {{
                background: rgba(0, 0, 0, 0.8);
                border: 2px solid #00ff88;
                border-radius: 15px;
                padding: 20px;
                box-shadow: 0 0 20px rgba(0, 255, 136, 0.3);
            }}
            
            .status-card h3 {{
                margin-bottom: 15px;
                color: #00ff88;
                border-bottom: 1px solid #00ff88;
                padding-bottom: 10px;
            }}
            
            .status-item {{
                display: flex;
                justify-content: space-between;
                margin: 10px 0;
                padding: 8px;
                background: rgba(0, 255, 136, 0.1);
                border-radius: 5px;
            }}
            
            .success {{
                color: #00ff88;
            }}
            
            .warning {{
                color: #ffaa00;
            }}
            
            .button {{
                background: linear-gradient(45deg, #00ff88, #00cc66);
                color: #000;
                border: none;
                padding: 15px 30px;
                border-radius: 25px;
                font-size: 1.1rem;
                font-weight: bold;
                cursor: pointer;
                margin: 10px;
                transition: all 0.3s ease;
                text-decoration: none;
                display: inline-block;
            }}
            
            .button:hover {{
                transform: translateY(-2px);
                box-shadow: 0 5px 15px rgba(0, 255, 136, 0.4);
            }}
            
            .footer {{
                text-align: center;
                margin-top: 40px;
                padding: 20px;
                border-top: 1px solid #00ff88;
                color: #666;
            }}
            
            @media (max-width: 768px) {{
                .logo {{
                    font-size: 2rem;
                }}
                
                .status-grid {{
                    grid-template-columns: 1fr;
                }}
                
                .status-item {{
                    flex-direction: column;
                    text-align: left;
                }}
            }}
        </style>
    </head>
    <body>
        <div class="container">
            <div class="header">
                <div class="logo">🚀 ARBTRONX</div>
                <div class="subtitle">Live Trading Dashboard - Successfully Deployed!</div>
            </div>
            
            <div class="status-grid">
                <div class="status-card">
                    <h3>🎯 Deployment Status</h3>
                    <div class="status-item">
                        <span>Platform:</span>
                        <span class="success">✅ RENDER.COM</span>
                    </div>
                    <div class="status-item">
                        <span>Health Check:</span>
                        <span class="success">✅ PASSING</span>
                    </div>
                    <div class="status-item">
                        <span>Application:</span>
                        <span class="success">✅ RUNNING</span>
                    </div>
                    <div class="status-item">
                        <span>Version:</span>
                        <span class="success">v1.0.0 (Lite)</span>
                    </div>
                </div>
                
                <div class="status-card">
                    <h3>👤 User Profile</h3>
                    <div class="status-item">
                        <span>Name:</span>
                        <span class="success">{user_name}</span>
                    </div>
                    <div class="status-item">
                        <span>Email:</span>
                        <span class="success">{user_email}</span>
                    </div>
                    <div class="status-item">
                        <span>User ID:</span>
                        <span class="success">{user_id}</span>
                    </div>
                    <div class="status-item">
                        <span>API Key:</span>
                        <span class="success">{masked_key}</span>
                    </div>
                </div>
                
                <div class="status-card">
                    <h3>💰 Trading System</h3>
                    <div class="status-item">
                        <span>Goal:</span>
                        <span class="success">$200 → $100,000</span>
                    </div>
                    <div class="status-item">
                        <span>Strategy:</span>
                        <span class="success">Grid Trading + Phase System</span>
                    </div>
                    <div class="status-item">
                        <span>Pairs:</span>
                        <span class="success">PEPE, FLOKI, DOGE, SHIB, SUI</span>
                    </div>
                    <div class="status-item">
                        <span>Mode:</span>
                        <span class="warning">⚡ LITE VERSION</span>
                    </div>
                </div>
                
                <div class="status-card">
                    <h3>🔧 Next Steps</h3>
                    <div class="status-item">
                        <span>Phase 1:</span>
                        <span class="success">✅ Deployment Complete</span>
                    </div>
                    <div class="status-item">
                        <span>Phase 2:</span>
                        <span class="warning">🔄 Add Trading Engine</span>
                    </div>
                    <div class="status-item">
                        <span>Phase 3:</span>
                        <span class="warning">🔄 Add Live Market Data</span>
                    </div>
                    <div class="status-item">
                        <span>Phase 4:</span>
                        <span class="warning">🔄 Enable Live Trading</span>
                    </div>
                </div>
            </div>
            
            <div style="text-align: center;">
                <a href="/health" class="button">🏥 Health Check</a>
                <a href="/status" class="button">📊 Status</a>
                <button class="button" onclick="window.location.reload()">🔄 Refresh</button>
            </div>
            
            <div class="footer">
                <p>🌐 Your ARBTRONX dashboard is now live on Render!</p>
                <p>📱 Mobile-responsive and ready for incremental feature deployment</p>
                <p>🔒 Secure environment variables configured</p>
                <p>⚡ Next: Deploy full trading bot features step by step</p>
            </div>
        </div>
    </body>
    </html>
    """
    
    return html_content

if __name__ == "__main__":
    port = int(os.environ.get("PORT", 8000))
    print("🚀 Starting ARBTRONX Live Dashboard (Lite Version)...")
    print("📱 Mobile-responsive interface ready")
    print("💰 Foundation for live trading system")
    print(f"🌐 Dashboard will be available at: http://0.0.0.0:{port}")
    print("=" * 80)
    uvicorn.run(app, host="0.0.0.0", port=port)
